requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.1
httpx==0.24.0
idna==3.4
kombu==5.2.4
liqpay-sdk-python3==1.0.3
postgres==4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
psycopg2-pool==1.1
pyTelegramBotAPI==4.12.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
